titomachine::emulator::devices

Trait MMIO

pub(crate) trait MMIO: Device {
    // Required methods
    fn read(&mut self, addr: usize) -> Result<i32, ()>;
    fn write(&mut self, addr: usize, value: i32) -> Result<(), ()>;
}

Memory Mapped IO: Any device that occupies memory addresses shall implement this trait.

Required Methods

fn read(&mut self, addr: usize) -> Result<i32, ()>

MMIO read. In implementation, address is relative to device offset, not global. So your first addr is always 0x0.

fn write(&mut self, addr: usize, value: i32) -> Result<(), ()>

MMIO write. In implementation, address is relative to device offset, not global. So your first addr is always 0x0.