Struct titomachine::gui::memoryview::MemoryView

pub(crate) struct MemoryView {
    view_cache_start: usize,
    view_cache_size: usize,
    view_cache: HashMap<usize, i32>,
    symbol_table: HashMap<usize, Vec<String>>,
    breakpoints: HashSet<usize>,
    pub is_playing: bool,
    pub cpu_pc: usize,
    pub cpu_sp: usize,
    pub cpu_fp: usize,
    pub start_code: usize,
    pub start_data: usize,
    pub start_stack: usize,
}

MemoryView is the UI component responsible for the memory viewer panel.

Fields
view_cache_start: usize

First address to cache

view_cache_size: usize

Number of addresses to cache

view_cache: HashMap<usize, i32>

Currently cached addresses

symbol_table: HashMap<usize, Vec<String>>

Multiple symbols may exist for an address, because of consts

breakpoints: HashSet<usize>

Set of addresses that contains a breakpoint.

is_playing: bool

Is the emulated machine both turned on and not paused

cpu_pc: usize

Memory view displays where PC, SP, and FP point to

cpu_sp: usize

Memory view displays where PC, SP, and FP point to

cpu_fp: usize

Memory view displays where PC, SP, and FP point to

start_code: usize

Code segment start address

start_data: usize

Data segment start address

start_stack: usize

Stack start address

Implementations

impl MemoryView

pub fn new() -> Self

pub fn reset(&mut self)

Reset needs to be called when loading a new program. Otherwise, old stuff like breakpoints may linger in the memory view despite being removed from the emulator.

pub fn get_view_cache_range(&self) -> Range<u32>

Range of addresses currently visible on screen

pub fn get_view_cache_start(&self) -> usize

First address currently visible on screen

pub fn set_view_cache(&mut self, range_start: usize, values: Vec<i32>)

Update memoryview’s cached addresses. range_start tells the first address. Use get_view_cache_range() to know which addresses memoryview wants.

pub fn set_symbol_table(&mut self, table: HashMap<String, i32>)

Give memoryview a copy of B91 symbol table.

pub fn jump_to_pc(&mut self)

Scroll view to PC location

fn get_segment_from_address(&self, address: usize) -> MemorySegment

Which segment does an address belong?

fn get_table_rows_fit(&self, ui: &Ui) -> usize

Determine how many table rows can fit on screen. May give one too many, but overflow should be hidden.

fn add_table_row( &mut self, config: &mut Config, sender: &Sender<CtrlMSG>, body: &mut TableBody<'_>, address: usize )

Add an address-entry-displaying row to the memory view table.

fn add_table_label( &self, row: &mut TableRow<'_, '_>, text: &str, font_color: Color32 )

Table Shortcut: Generic label

fn add_table_address( &mut self, config: &Config, sender: &Sender<CtrlMSG>, row: &mut TableRow<'_, '_>, address: usize, font_color: Color32 )

Table Shortcut: Address column

fn add_table_value( &self, config: &mut Config, row: &mut TableRow<'_, '_>, value: i32, font_color: Color32 )

Table Shortcut: Value column

fn add_table_disassembly( &self, row: &mut TableRow<'_, '_>, value: i32, font_color: Color32 )

Table Shortcut: Disassembly column

fn add_table_pointers( &self, row: &mut TableRow<'_, '_>, address: usize, font_color: Color32 )

Add a column that shows if PC, SP, FP, or any symbols point to this address

fn add_table_heading(&self, header: &mut TableRow<'_, '_>, title: &str)

Table Shortcut: Header column